shuffle things around
